Удаление предупреждений PVS-Studio

+++ source/TacticsVictory/src/Game/Objects/Ammo/Rocket/Rocket.cpp
@@ -304,7 +304,7 @@ void Rocket::UpdateOn()
             VerifyOnIntersectionTerrain();
         }
 
-        FuncUpdate func = funcs[state];
+        FuncUpdate func = funcs[static_cast<int>(state)];
         (this->*func)();
 
         CalculateAnimate();
